Extract AboutModal to new file
